Enable sqrtf() x86 assembly code
This is a forward porting of optimization for OTC bugzilla #13692.
Below is the description of Honeycomb behavior.

RenderScript Demo Balls is slower on Redridge than on ARM devices.
The FPS on RedRidge is 37 while most ARM devices can reach 60.
This is cpu intensive application. The SEP data shows that sqrtf() in
libm.so costs about 20% time.

We can see that sqrtf() on x86 has an assembly version in i387/e_sqrtf.S.
And we can get about 19% improvement for Balls by enabling the assembly
version.

RsBalls on dv10 with ICS currently only reach 17 fps. This patch can improve
it to 19 fps.

This change modified makefile to enable the assembly version.

Change-Id: I04a712e3949adc8e90a3dddc7eac6b8fcfe43a4f