Add mmap64()
This adds mmap64() to bionic so that it is possible to have
large offset passed to kernel. However, the syscall mechanism
only passes 32-bit number to kernel. So effectively, the
largest offset that can be passed is about 43 bits (since
offset is signed, and the number passed to kernel is number
of pages (page size == 4K => 12 bits)).

For: AXIA-1343
Change-Id: I4f7699dfd4d1c7a2bd46fdd646f0c4a90dc4eb72
Depends-Change-Id: Idb207d912ec50dffa51369c6581e9255a4ca15ae
Signed-off-by: Daniel Leung <daniel.leung@intel.com>