Add mmap64()
This adds mmap64() to bionic so that it is possible to have
large offset passed to kernel. However, the syscall mechanism
only passes 32-bit number to kernel. So effectively, the
largest offset that can be passed is about 43 bits (since
offset is signed, and the number passed to kernel is number
of pages (page size == 4K => 12 bits)).

Upstream change:
https://android-review.googlesource.com/#/c/65726/

Category: Device Enablement
Domain: AOSP-bionic
Source: Internal
Upstream-candidate: Yes
Change-Id: Ib54f4e9b54acb6ef8b0324f3b89c9bc810b07281
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Depends-Change-Id: I63692629d3a62b7aee21202b67241904b1cc9789
Depends-Change-Id: Idb0c3d3bb98f22ea7d78043b786e931613377b02